+++ pkg/providers/reportinggroups/provider.go
@@ -58,6 +58,7 @@ func (p *Subprovider) FrameworkResources() []func() resource.Resource {
 func (p *Subprovider) FrameworkDataSources() []func() datasource.DataSource {
 	return []func() datasource.DataSource{
 		NewCPCodeDataSource,
+		NewCPCodesDataSource,
 		NewReportingGroupDataSource,
 		NewReportingGroupsDataSource,
 	}