+++ pkg/providers/reportinggroups/testdata/TestDataCPCodes/cp_codes.tf
@@ -1,0 +1,6 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_reportinggroups_cp_codes" "test" {}
+