+++ pkg/providers/reportinggroups/testdata/TestDataCPCodes/cp_codes_product_id_only.tf
@@ -1,0 +1,8 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_reportinggroups_cp_codes" "test" {
+  product_id = "Site_Accel::Site_Accel"
+}
+