+++ pkg/providers/reportinggroups/testdata/TestDataCPCodes/cp_codes_with_filters.tf
@@ -1,0 +1,11 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_reportinggroups_cp_codes" "test" {
+  contract_id  = "1-2ABCDE"
+  group_id     = "grp_67890"
+  product_id   = "Web_Exp::Ion_Na"
+  cp_code_name = "My CP Code"
+}
+